Rename BloggingPromptsOnboardingViewModel UiState and Action LiveData
