+++ WordPress/src/main/java/org/wordpress/android/ui/bloggingprompts/onboarding/BloggingPromptsOnboardingAction.kt
@@ -1,0 +1,7 @@
+package org.wordpress.android.ui.bloggingprompts.onboarding
+
+import org.wordpress.android.models.bloggingprompts.BloggingPrompt
+
+sealed class BloggingPromptsOnboardingAction {
+    data class OpenEditor(val bloggingPrompt: BloggingPrompt) : BloggingPromptsOnboardingAction()
+}
