+++ WordPress/src/main/java/org/wordpress/android/ui/bloggingprompts/onboarding/BloggingPromptsOnboardingUiState.kt
@@ -1,3 +1,3 @@
 package org.wordpress.android.ui.bloggingprompts.onboarding
 
-sealed class UiState
+sealed class BloggingPromptsOnboardingUiState
